requirements.txt
discord.py
elevenlabs
PyNaCl
ffmpeg
